+++ src/webviews/settingsWebview.ts
@@ -99,7 +99,7 @@ export class SettingsWebview extends WebviewBase {
 			scope: 'user',
 			scopes: scopes,
 		};
-		return `<script type="text/javascript" nonce="Z2l0bGVucy1ib290c3RyYXA=">window.bootstrap = ${JSON.stringify(
+		return `<script type="text/javascript" nonce="#{cspNonce}">window.bootstrap = ${JSON.stringify(
 			bootstrap,
 		)};</script>`;
 	}